requirements.txt
requests
beautifulsoup4
python-dotenv
duckdb==0.10.0
pandas
dbt-core
dbt-duckdb 